+++ src/mediawiki.ts
@@ -15,7 +15,6 @@ import {isUnderscore} from '@bhsd/cm-util';
 import {commonHtmlAttrs, htmlAttrs, extAttrs} from 'wikiparser-node/dist/util/sharable.mjs';
 import {htmlTags, tokens} from './config.js';
 import {
-	hoverSelector,
 	extCompletion,
 	isWMF,
 	mwPrefix,
@@ -31,6 +30,7 @@ import {
 	useUnderscore,
 	getHighlightExtension,
 } from './util.js';
+import {hoverStyle} from './hover.js';
 import type {
 	TagStyle,
 } from '@codemirror/language';
@@ -755,28 +755,10 @@ const wikiTheme = /* @__PURE__ */ EditorView.theme({
 	[`${mwPrefix}tag-ref`]: {
 		backgroundColor: 'var(--cm-ref)',
 	},
-
-	// hover tooltip and signature tooltip
-	'.cm-tooltip-hover': {
-		maxHeight: '60vh',
-		overflow: 'hidden auto',
-	},
-	[hoverSelector]: {
-		padding: '2px 5px',
-		width: 'max-content',
-		maxWidth: '60vw',
-		'& *': {
-			marginTop: '0!important',
-			marginBottom: '0!important',
-		},
-		'&>div': {
-			fontSize: '90%',
-			lineHeight: 1.4,
-		},
-	},
 	'.cm-completionIcon-redirect:after': {
 		content: '"↳"',
 	},
+	...hoverStyle,
 });
 
 /**