+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 
 **Added**
 
-- The [openLinks](./README.md#openlinks) extension and the [autocompletion](./README.md#autocompletion) extension now also supports internal links starting with `../` on a MediaWiki site
+- The [openLinks](./README.md#openlinks) extension and the [autocompletion](./README.md#autocompletion) extension now supports internal links starting with `../` in the MediaWiki mode on a MediaWiki site
+- The [openLinks](./README.md#openlinks) extension now also supports internal links containing HTML comments in the MediaWiki mode on a MediaWiki site
 
 ## 3.13.0